+++ Sources/SwiftUIKit/_Deprecated/FlippableView.swift
@@ -1,0 +1,6 @@
+#if os(iOS)
+import SwiftUI
+
+@available(*, deprecated, renamed: "FlipView")
+public typealias FlippableView = FlipView
+#endif